fix(haetae): harden FFT fixed-point widths
Make the HAETAE key-rejection FFT width model explicit and portable end-to-end.

- Store FFT real/imaginary components as int32_t under the documented B8 <= 859963392 proof.
- Evaluate Q16 multiplication and butterfly intermediates in int64_t before proven int32_t stores.
- Return squared magnitudes as uint64_t and keep singular-value sums, best-m sorting, and accumulation in uint64_t.
- Replace signed-right-shift Q16 arithmetic with the shared portable floor-division helper.
- Use branchless uint64_t compare/swap under a compile-time <2^63 bound.
- Tie the proof to ETA=1, d<=1, N=256, eight FFT stages, and current maximum parameter dimensions with static assertions.
- Preserve the intended final rounding as (result + UINT64_C(32)) >> 6.

Validation before merge: HAETAE unit plus current and legacy 120/180/260 KATs passed; ASan+UBSan HAETAE unit passed; Windows/MSVC, Ubuntu, and macOS builds succeeded in PR CI.

+++ src/DigitalSignature/HAETAE/haetae_polyvec.h
@@ -75,7 +75,10 @@ void crypto_haetae_polyvecm_pointwise_acc_montgomery(crypto_haetae_poly *w,
 
 void crypto_haetae_polyvecm_ntt(crypto_haetae_polyvecm *x, const crypto_haetae_parameters *parameters);
 
-long long crypto_haetae_polyvecmk_sqsing_value(const crypto_haetae_polyvecm *s1, const crypto_haetae_polyveck *s2, const crypto_haetae_parameters *parameters);
+uint64_t crypto_haetae_polyvecmk_sqsing_value(
+    const crypto_haetae_polyvecm *s1,
+    const crypto_haetae_polyveck *s2,
+    const crypto_haetae_parameters *parameters);
 
 long long crypto_haetae_polyvecmk_sing_value(const crypto_haetae_polyvecm *s1, const crypto_haetae_polyveck *s2, const crypto_haetae_parameters *parameters);
